x_deep_agents/research_agent/pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.4.12",
    "langchain-anthropic>=1.4.0",
    "markdownify>=1.2.2",
    "python-dotenv>=1.2.2",
    "tavily>=1.1.0",
]
